[IMPROV] Feed invalid IMPROV messages to console input again
